Move default org logic from placeholder to httpstate package
