+++ pkg/cmd/pulumi/packagecmd/package_publish.go
@@ -27,9 +27,9 @@ import (
 	"github.com/blang/semver"
 	"github.com/pulumi/pulumi/pkg/v3/backend"
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
+	"github.com/pulumi/pulumi/pkg/v3/backend/httpstate"
 	cmdBackend "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/backend"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/schema"
-	"github.com/pulumi/pulumi/pkg/v3/placeholder"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/env"
@@ -88,7 +88,7 @@ func newPackagePublishCmd() *cobra.Command {
 		Hidden: !env.Experimental.Value(),
 		RunE: func(cmd *cobra.Command, cliArgs []string) error {
 			ctx := cmd.Context()
-			pkgPublishCmd.defaultOrg = placeholder.GetDefaultOrg
+			pkgPublishCmd.defaultOrg = httpstate.GetDefaultOrg
 			pkgPublishCmd.extractSchema = SchemaFromSchemaSource
 			return pkgPublishCmd.Run(ctx, args, cliArgs[0], cliArgs[1:])
 		},
